[SDL] Do the GetTickCountWord() math unsigned
Related changes:
* [internal] Fix a Uint32-to-WORD casting warning

+++ sdl/timer.cpp
@@ -6,7 +6,7 @@ QWORD perf_freq;
 
 WORD GetTickCountWord()
 {
-	return SDL_GetTicks() * 13125LL / (11*65535);
+	return (WORD)(SDL_GetTicks() * 13125uLL / (11*65535));
 }
 
 int OpenTimer()
